requirements.txt
fastapi
uvicorn
httpx
pydantic
pytest
pytest-cov
pytest-mock
pytest-asyncio
python-jose
motor
passlib[bcrypt]
